fix(keymap): ⌘,/Ctrl+, — 쉼표 키가 2단 코드 구분자로 잘리던 결함(split_seq) · feat(log): ext 층(마스크 u64) + 확장 매니저 상세 추적(다운로드 바이트/시간/속도 · 검증 · 보관/배치 폴더 · 기록) (사용자 요청 09-17)

+++ crates/nexa-sql/src/log_win.rs
@@ -123,7 +123,7 @@ pub(crate) struct LogWin {
     dev: bool,
     dev_switch: Switch,
     /// 상세 층 마스크(메뉴 체크 표시용 · 설정 `log.dev_layers`).
-    dev_mask: u32,
+    dev_mask: u64,
 }
 
 impl LogWin {
@@ -190,7 +190,7 @@ impl LogWin {
     }
 
     /// 상세 층 마스크(메뉴 체크 표시).
-    pub(crate) fn set_dev_mask(&mut self, mask: u32) {
+    pub(crate) fn set_dev_mask(&mut self, mask: u64) {
         self.dev_mask = mask;
     }
